pyproject.toml
[project]
name = "gosend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "aiomysql>=0.2.0",
    "alembic>=1.15.1",
    "decouple>=0.0.7",
    "fastapi[standard]>=0.115.11",
    "greenlet>=3.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "redis>=5.2.1",
    "setuptools>=76.1.0",
    "sqlalchemy>=2.0.38",
]
